requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
anyio==3.6.1
arrow==1.2.3
asgiref==3.5.2
astor==0.8.1
asynctest==0.13.0
attrs==22.1.0
autoflake==1.5.3
bandit==1.7.4
binaryornot==0.4.4
black==22.8.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.14
cffi==1.15.1
cfgv==3.3.1
chardet==5.0.0
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
cookiecutter==1.7.3
coverage==6.4.4
crashtest==0.3.1
darglint==1.8.1
dictdiffer==0.9.0
distlib==0.3.6
docutils==0.19
dulwich==0.20.46
eradicate==2.1.0
fastapi==0.75.2
fastapi-template==3.3.8
filelock==3.8.0
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.4.0
flake8-bugbear==22.9.11
flake8-commas==2.1.0
flake8-comprehensions==3.10.0
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-eradicate==1.3.0
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
gitdb==4.0.9
GitPython==3.1.27
h11==0.12.0
html5lib==1.1
httpcore==0.14.7
httptools==0.3.0
httpx==0.22.0
identify==2.5.5
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
iso8601==1.0.2
isort==5.10.1
jaraco.classes==3.2.2
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.16.0
keyring==23.9.1
lockfile==0.12.2
MarkupSafe==2.1.1
mccabe==0.6.1
more-itertools==8.14.0
msgpack==1.0.4
multidict==6.0.2
mypy==0.961
mypy-extensions==0.4.3
nest-asyncio==1.5.5
nodeenv==1.7.0
packaging==21.3
pathspec==0.10.1
pbr==5.10.0
pep8-naming==0.12.1
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.2.0
poetry-core==1.1.0
poetry-plugin-export==1.0.6
poyo==0.5.0
pre-commit==2.20.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.13.0
pylev==1.4.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
pyrsistent==0.18.1
pytest==7.1.3
pytest-cov==3.0.0
pytest-env==0.6.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
restructuredtext-lint==1.4.0
rfc3986==1.5.0
shellingham==1.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
starlette==0.17.1
stevedore==4.0.0
termcolor==1.1.0
text-unidecode==1.3
tokenize-rt==4.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tortoise-orm==0.19.2
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.17.5
uvloop==0.17.0
virtualenv==20.16.5
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
wemake-python-styleguide==0.16.1
xattr==0.9.9
yarl==1.8.1
yesqa==1.4.0
zipp==3.8.1
